:wrench: refactor RelayRoot into App
